refactor: Add hideOnClickOutside feature to reuse code (#240)
* remove console log

* refactor: Add hideOnClickOutside feature to reuse code

This change abstracts the code repetition in three different files and promote the DRY principle by adding the hideOnClickOutside feature to the handleClickOutside function. The feature is designed to hide an element when clicking outside of its area.

* refactor: hooks

---------
